feat(site): server-render a title, H1, and pitch for notes, hikes, ships, and stars
The four csr-only apps returned an empty shell to crawlers, curl, and
llms.txt readers, including /app/notes, the homepage's primary CTA. Each
page now renders its head, H1, and one paragraph on the server and mounts
the browser-only app (d3 graph, MapLibre maps) from a sibling component
behind an if-browser guard. SSR shell tests render each page with
svelte/server; the SvelteKit navigation and stores virtual modules get
vitest stubs like the existing environment stub.
