+++ projects/monolith/frontend/BUILD
@@ -207,6 +207,10 @@ js_library(
         # Resolved by the $app/environment alias in vitest.config.js; lives
         # outside src/ so it never reaches the app bundle.
         "test/app-environment-stub.js",
+        # Client-only page components imported by SSR shell tests reference
+        # these SvelteKit virtual modules even though they never mount.
+        "test/app-navigation-stub.js",
+        "test/app-stores-stub.js",
         # Resolved by the $env/dynamic/private alias in vitest.config.js; same
         # rationale: lives outside src/ so it never reaches the app bundle.
         "test/env-dynamic-private-stub.js",
