+++ projects/monolith/frontend/src/routes/public/app/hikes/+page.js
@@ -1,5 +1,3 @@
-// MapLibre needs window/WebGL, so render the page client-side only. The
-// +page.server.js load still runs server-side regardless, so the walks data
-// stays SSR-sourced (the browser never touches /api/hikes/*). Same pattern as
-// /app/ships.
-export const ssr = false;
+// Render the static page pitch on the server. The MapLibre application mounts
+// in the browser from HikesApp.svelte.
+export const ssr = true;
