+++ projects/monolith/frontend/src/routes/public/app/stars/+page.js
@@ -1,5 +1,3 @@
-// MapLibre needs window/WebGL, so render the page client-side only. The
-// +page.server.js load still runs server-side regardless, so the sites data
-// stays SSR-sourced (the browser never touches /api/stars/*). Same pattern as
-// /app/hikes and /app/ships.
-export const ssr = false;
+// Render the static page pitch on the server. The MapLibre application mounts
+// in the browser from StarsApp.svelte.
+export const ssr = true;
