+++ projects/monolith/frontend/test/app-navigation-stub.js
@@ -1,0 +1,3 @@
+export async function goto() {}
+
+export async function invalidateAll() {}
