+++ projects/monolith/frontend/test/app-stores-stub.js
@@ -1,0 +1,10 @@
+const value = {
+  url: new URL("https://example.test/"),
+};
+
+export const page = {
+  subscribe(run) {
+    run(value);
+    return () => {};
+  },
+};
